review: extract BlockChain interface to chain/internal/common
