+++ chain/aptos/aptos_chain.go
@@ -53,7 +53,7 @@ func (c Chain) IsNetworkType(networkType chainsel.NetworkType) bool {
 	return chaincommon.ChainMetadata{Selector: c.Selector}.IsNetworkType(networkType)
 }
 
-func (c Chain) ReadOnly() (any, error) {
+func (c Chain) ReadOnly() (chaincommon.BlockChain, error) {
 	signer, err := aptoslib.NewSecp256k1Account()
 	if err != nil {
 		return nil, fmt.Errorf("unable to generate private key for read-only chain %v: %w", c, err)
