+++ chain/solana/solana_chain.go
@@ -220,7 +220,7 @@ func (c Chain) IsNetworkType(networkType chainsel.NetworkType) bool {
 	return chaincommon.ChainMetadata{Selector: c.Selector}.IsNetworkType(networkType)
 }
 
-func (c Chain) ReadOnly() (any, error) {
+func (c Chain) ReadOnly() (chaincommon.BlockChain, error) {
 	privateKey, err := sollib.NewRandomPrivateKey()
 	if err != nil {
 		return nil, fmt.Errorf("failed to generate private key for read-only chain %v: %w", c.Selector, err)
