+++ engine/test/internal/testutils/testutils.go
@@ -32,4 +32,4 @@ func (c *StubChain) NetworkType() (chainsel.NetworkType, error) {
 func (c *StubChain) IsNetworkType(networkType chainsel.NetworkType) bool {
 	return networkType == chainsel.NetworkTypeTestnet
 }
-func (c *StubChain) ReadOnly() any { return c }
+func (c *StubChain) ReadOnly() (fchain.BlockChain, error) { return c, nil }
